// src/api/nexuscore/model/onlineMember.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Nexuscore
 * The backend for all Everthorn internal services
 * OpenAPI spec version: 1.0.0
 */

export interface OnlineMember {
  /** The ThornyID of the user */
  thorny_id: number;
  /** The Discord ID of the user */
  user_id: number;
  /** The date and time when the session started */
  session: string;
  /** The username of the user */
  username: string;
  /** The gamertag of the user */
  whitelist: string;
  /**
     * The last in-game location of the user
     * @minItems 3
     * @maxItems 3
     */
  location: [number, number, number];
  /** The last in-game dimension the user was in */
  dimension: string;
  /** Whether the user should be hidden on the Live Map */
  hidden: boolean;
  xuid: string | null;
}